test: Refactor to rely on fixtures setup
