+++ lib/Serverless.test.js
@@ -7,7 +7,6 @@ const path = require('path');
 const YAML = require('js-yaml');
 const sinon = require('sinon');
 const BbPromise = require('bluebird');
-const fse = require('fs-extra');
 
 const YamlParser = require('../lib/classes/YamlParser');
 const PluginManager = require('../lib/classes/PluginManager');
@@ -287,52 +286,46 @@ describe('Serverless', () => {
 });
 
 describe('Serverless [new tests]', () => {
-  describe('When local version available', () => {
-    before(() => {
-      const fixturePath = path.resolve(fixtures.map.aws, 'node_modules/serverless');
-      const fixtureModulePath = path.resolve(fixturePath, 'index.js');
-      return fse
-        .copy(path.resolve(fixtures.map.aws, 'node_modules-fixture/serverless'), fixturePath)
-        .then(() => fse.readFile(fixtureModulePath))
-        .then(content =>
-          fse.writeFile(
-            fixtureModulePath,
-            String(content).replace(
-              '$SERVERLESS_PATH',
-              JSON.stringify(path.resolve(__dirname, '../'))
-            )
-          )
-        );
-    });
-    after(() => fse.remove(path.resolve(fixtures.map.aws, 'node_modules')));
+  before(() => fixtures.setup('locallyInstalledServerless'));
+  after(fixtures.cleanup);
 
+  describe('When local version available', () => {
     describe('When running global version', () => {
       it('Should report deprecation notice when "enableLocalInstallationFallback" not set', () =>
-        runServerless({ cwd: fixtures.map.aws, cliArgs: ['-v'] }).then(({ serverless }) => {
-          expect(Array.from(serverless.triggeredDeprecations)).to.deep.equal([
-            'LOCAL_INSTALLATION_FALLBACK',
-          ]);
-          expect(serverless.invokedInstance).to.not.exist;
-        }));
-      it('Should run without notice when "enableLocalInstallationFallback" set to false', () =>
-        fixtures.extend('aws', { enableLocalInstallationFallback: false }).then(fixturePath =>
-          runServerless({ cwd: fixturePath, cliArgs: ['-v'] }).then(({ serverless }) => {
-            expect(Array.from(serverless.triggeredDeprecations)).to.deep.equal([]);
+        runServerless({ cwd: fixtures.map.locallyInstalledServerless, cliArgs: ['-v'] }).then(
+          ({ serverless }) => {
+            expect(Array.from(serverless.triggeredDeprecations)).to.deep.equal([
+              'LOCAL_INSTALLATION_FALLBACK',
+            ]);
             expect(serverless.invokedInstance).to.not.exist;
-          })
+          }
         ));
+      it('Should run without notice when "enableLocalInstallationFallback" set to false', () =>
+        fixtures
+          .extend('locallyInstalledServerless', { enableLocalInstallationFallback: false })
+          .then(fixturePath =>
+            runServerless({ cwd: fixturePath, cliArgs: ['-v'] }).then(({ serverless }) => {
+              expect(Array.from(serverless.triggeredDeprecations)).to.deep.equal([]);
+              expect(serverless.invokedInstance).to.not.exist;
+            })
+          ));
       it('Should fallback to local version when "enableLocalInstallationFallback" set to true', () =>
-        fixtures.extend('aws', { enableLocalInstallationFallback: true }).then(fixturePath =>
-          runServerless({ cwd: fixturePath, cliArgs: ['-v'] }).then(({ serverless }) => {
-            expect(Array.from(serverless.triggeredDeprecations)).to.deep.equal([]);
-            expect(serverless.invokedInstance.isLocalStub).to.be.true;
-          })
-        ));
+        fixtures
+          .extend('locallyInstalledServerless', { enableLocalInstallationFallback: true })
+          .then(fixturePath =>
+            runServerless({ cwd: fixturePath, cliArgs: ['-v'] }).then(({ serverless }) => {
+              expect(Array.from(serverless.triggeredDeprecations)).to.deep.equal([]);
+              expect(serverless.invokedInstance.isLocalStub).to.be.true;
+            })
+          ));
     });
     describe('When running local version', () => {
       it('Should run without notice', () =>
         runServerless({
-          serverlessPath: path.resolve(fixtures.map.aws, 'node_modules/serverless'),
+          serverlessPath: path.resolve(
+            fixtures.map.locallyInstalledServerless,
+            'node_modules/serverless'
+          ),
           cwd: fixtures.map.aws,
           cliArgs: ['-v'],
         }).then(({ serverless }) => {
