+++ test/fixtures/locallyInstalledServerless/node_modules-fixture/serverless/index.js
@@ -9,6 +9,7 @@ module.exports = class LocalServerless extends Serverless {
     this.isLocalStub = true;
   }
   eventuallyFallbackToLocal() {
+    this.isLocallyInstalled = true;
     return Promise.resolve();
   }
 };
