+++ test/fixtures/locallyInstalledServerless/node_modules-fixture/serverless/lib/utils/analytics/areDisabled.js
