+++ test/fixtures/locallyInstalledServerless/serverless.yml
@@ -1,0 +1,7 @@
+service: service
+
+configValidationMode: error
+frameworkVersion: '*'
+
+provider:
+  name: aws
